requirements.txt
ollama
pillow
pyperclip
keyboard
pytesseract